+++ packages/vike-react-apollo/src/integration/Transport.tsx
@@ -3,18 +3,34 @@ import { buildManualDataTransport } from '@apollo/client-react-streaming/manual-
 import { useStream } from 'react-streaming'
 import { renderToString } from 'react-dom/server'
 import React from 'react'
+import { assert } from '../utils/assert.js'
+import { usePageContext } from 'vike-react/usePageContext'
 
 export const WrappedApolloProvider = WrapApolloProvider(
   buildManualDataTransport({
     useInsertHtml() {
       const stream = useStream()
-      if (!stream) {
+      const pageContext = usePageContext()
+      if (globalThis.__VIKE__IS_CLIENT) {
+        // Apollo only calls useInsertHtml() on the server
+        assert(false as boolean) // `as boolean` to avoid TypeScript complaining
+        // Enable tree-shaking
         return () => {}
       }
+      assert(!pageContext.isClientSide)
+      assert(stream)
       return (callback: () => React.ReactNode) => {
         stream.injectToStream(
           // https://github.com/apollographql/apollo-client-nextjs/issues/325
-          (async () => renderToString(await callback()))(),
+          (async () => {
+            let html = renderToString(await callback())
+            const nonce = pageContext.cspNonce
+            if (nonce) {
+              // No need to escape the injected HTML — see https://github.com/vikejs/vike/blob/36201ddad5f5b527b244b24d548014ec86c204e4/packages/vike/src/server/runtime/renderPageServer/csp.ts#L45
+              html = html.replace(/<script(\s|>)/g, `<script nonce="${nonce}"$1`)
+            }
+            return html
+          })(),
         )
       }
     },