+++ packages/vike-react-zustand/src/utils/assert.ts
@@ -2,9 +2,7 @@ export { assert, assertUsage }
 
 function assert(condition: unknown): asserts condition {
   if (condition) return
-  throw new Error(
-    "You stumbled upon a bug in vike-react-zustand's source code. Reach out on GitHub and we will fix the bug.",
-  )
+  throw new Error('You stumbled upon a vike-react-zustand bug, reach out on GitHub.')
 }
 
 function assertUsage(condition: unknown, message: string): asserts condition {