+++ packages/vike-react/src/helpers/clientOnly.tsx
@@ -1,3 +1,6 @@
+// TODO/soon: add deprecation warning in favor of <ClientOnly>
+// TO-DO/breaking-change: remove it
+
 export { clientOnly }
 
 import React, {